+++ apps/app/components/views/form.tsx
@@ -6,14 +6,7 @@ import useSWR from "swr";
 
 import { useForm } from "react-hook-form";
 // ui
-import {
-  Avatar,
-  Input,
-  MultiLevelDropdown,
-  PrimaryButton,
-  SecondaryButton,
-  TextArea,
-} from "components/ui";
+import { Avatar, Input, PrimaryButton, SecondaryButton, TextArea } from "components/ui";
 // types
 import { IView } from "types";
 // constant
@@ -23,11 +16,12 @@ import { getStatesList } from "helpers/state.helper";
 // services
 import stateService from "services/state.service";
 import projectService from "services/project.service";
+// components
+import { SelectFilters } from "components/views";
 // icons
 import { getStateGroupIcon } from "components/icons";
 import { getPriorityIcon } from "components/icons/priority-icon";
 // components
-import { PRIORITIES } from "constants/project";
 
 type Props = {
   handleFormSubmit: (values: IView) => Promise<void>;
@@ -139,8 +133,8 @@ export const ViewForm: React.FC<Props> = ({
             />
           </div>
           <div>
-            <MultiLevelDropdown
-              label="Filters"
+            <SelectFilters
+              filters={filters}
               onSelect={(option) => {
                 const key = option.key as keyof typeof filters;
 
@@ -156,72 +150,6 @@ export const ViewForm: React.FC<Props> = ({
                   });
                 }
               }}
-              direction="right"
-              options={[
-                {
-                  id: "priority",
-                  label: "Priority",
-                  value: PRIORITIES,
-                  children: [
-                    ...PRIORITIES.map((priority) => ({
-                      id: priority ?? "none",
-                      label: (
-                        <div className="flex items-center gap-2">
-                          {getPriorityIcon(priority)} {priority ?? "None"}
-                        </div>
-                      ),
-                      value: {
-                        key: "priority",
-                        value: priority,
-                      },
-                      selected: filters?.priority?.includes(priority ?? "none"),
-                    })),
-                  ],
-                },
-                {
-                  id: "state",
-                  label: "State",
-                  value: statesList,
-                  children: [
-                    ...statesList.map((state) => ({
-                      id: state.id,
-                      label: (
-                        <div className="flex items-center gap-2">
-                          {getStateGroupIcon(state.group, "16", "16", state.color)} {state.name}
-                        </div>
-                      ),
-                      value: {
-                        key: "state",
-                        value: state.id,
-                      },
-                      selected: filters?.state?.includes(state.id),
-                    })),
-                  ],
-                },
-                {
-                  id: "assignee",
-                  label: "Assignee",
-                  value: members,
-                  children: [
-                    ...(members?.map((member) => ({
-                      id: member.member.id,
-                      label: (
-                        <div className="flex items-center gap-2">
-                          <Avatar user={member.member} />
-                          {member.member.first_name && member.member.first_name !== ""
-                            ? member.member.first_name
-                            : member.member.email}
-                        </div>
-                      ),
-                      value: {
-                        key: "assignees",
-                        value: member.member.id,
-                      },
-                      selected: filters?.assignees?.includes(member.member.id),
-                    })) ?? []),
-                  ],
-                },
-              ]}
             />
           </div>
           <div>
@@ -257,8 +185,9 @@ export const ViewForm: React.FC<Props> = ({
                 else if (queryKey === "assignees")
                   return (
                     <div className="flex gap-3" key={key}>
-                      {filters.assignees?.map((assigneeID) => {
-                        const member = members?.find((member) => member.member.id === assigneeID);
+                      {filters.assignees?.map((assigneeId) => {
+                        const member = members?.find((member) => member.member.id === assigneeId);
+
                         if (!member) return null;
                         return (
                           <div className="flex items-center gap-2 text-xs" key={member.member.id}>
