+++ apps/app/components/views/index.ts
@@ -1,3 +1,4 @@
 export * from "./delete-view-modal";
 export * from "./form";
 export * from "./modal";
+export * from "./select-filters";
